fix warnings

+++ src/NetBox/WinSCPDialogs.cpp
@@ -7302,7 +7302,7 @@ bool TSynchronizeChecklistDialog::Key(TFarDialogItem * Item, long KeyCode)
         Redraw();
         UpdateControls();
         if ((KeyCode == KEY_INS) &&
-            (Index < ListBox->GetItems()->GetCount() - 1))
+            (Index < static_cast<int>(ListBox->GetItems()->GetCount()) - 1))
         {
           ListBox->GetItems()->SetSelected(Index + 1);
         }
@@ -8162,7 +8162,7 @@ void TQueueDialog::RefreshQueue()
     size_t TopIndex = QueueListBox->GetItems()->GetTopIndex();
     size_t Index = TopIndex;
 
-    int ILine = 0;
+    size_t ILine = 0;
     while ((Index > ILine) &&
            (QueueListBox->GetItems()->GetObject(Index) ==
               QueueListBox->GetItems()->GetObject(Index - ILine - 1)))
